add download_all_AOML, mv module to files_XBT.jl
