+++ src/OceanRobots.jl
@@ -8,6 +8,7 @@ import Base: read
 include("types.jl")
 include("thredds_servers.jl")
 include("files.jl")
+include("files_XBT.jl")
 include("example_GOM.jl")
 
 export GDP, GDP_CloudDrift, NOAA, GliderFiles, ArgoFiles, OceanSites, OceanOPS, CCHDO, XBT
